Sergeev

See merge request mobile/saver-android!6

+++ app/src/main/java/digital/fact/saver/domain/models/Source.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "SOURCES")
 data class Source(
     @PrimaryKey(autoGenerate = true)
-    val _id: Int = 0,
+    val _id: Long = 0,
     val name: String,
     val type: Int,
     val start_sum: Long = 0,
